backend/requirements.txt
python-multipart
fastapi
uvicorn
requests
load-dotenv
astropy
pandas
astroquery
pillow
scipy